refactor: 클래스 및 패키지명 변경
