+++ src/main/java/roomescape/service/booking/reservation/module/PaymentService.java
@@ -3,18 +3,18 @@ package roomescape.service.booking.reservation.module;
 import org.springframework.stereotype.Service;
 import roomescape.dto.payment.PaymentRequest;
 import roomescape.dto.payment.PaymentResponse;
-import roomescape.infrastructure.payment.TossPaymentClient;
+import roomescape.infrastructure.tosspayments.TossPaymentsClient;
 
 @Service
 public class PaymentService {
 
-    private final TossPaymentClient tossPaymentClient;
+    private final TossPaymentsClient tossPaymentsClient;
 
-    public PaymentService(TossPaymentClient tossPaymentClient) {
-        this.tossPaymentClient = tossPaymentClient;
+    public PaymentService(TossPaymentsClient tossPaymentsClient) {
+        this.tossPaymentsClient = tossPaymentsClient;
     }
 
     public PaymentResponse pay(PaymentRequest paymentRequest) {
-       return tossPaymentClient.requestPayment(paymentRequest);
+       return tossPaymentsClient.requestPayment(paymentRequest);
     }
 }
